requirements.txt
alembic==1.7.7
atomicwrites==1.4.0
attrs==21.4.0
black==22.3.0
click==8.1.2
colorama==0.4.4
dnspython==2.2.1
email-validator==1.1.3
flake8==4.0.1
Flask==2.1.1
Flask-Login==0.6.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.1.1
python-dotenv==0.20.0
SQLAlchemy==1.4.35
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.1.1
WTForms==3.0.1
zipp==3.8.0
